+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-admin-bar/class-wpcom-admin-bar.php
@@ -20,24 +20,37 @@ class WPCOM_Admin_Bar extends \WP_Admin_Bar {
 		/**
 		 * Site menu
 		 */
-		'wp-admin/'                            => 'https://wordpress.com/home/%home_url%',
-		'wp-admin/plugins.php'                 => 'https://wordpress.com/plugins/%home_url%',
-		'wp-admin/themes.php'                  => 'https://wordpress.com/themes/%home_url%',
+		'wp-admin/'              => 'https://wordpress.com/home/%home_url%',
+		'wp-admin/plugins.php'   => 'https://wordpress.com/plugins/%home_url%',
+		'wp-admin/themes.php'    => 'https://wordpress.com/themes/%home_url%',
 
 		/**
 		 * +New menu
 		 */
-		'wp-admin/post-new.php'                => '/wp-admin/post-new.php?post_type=post&calypsoify=1',
-		'wp-admin/media-new.php'               => 'https://wordpress.com/media/%home_url%',
-		'wp-admin/post-new.php?post_type=page' => 'https://wordpress.com/page/%home_url%',
-		'wp-admin/user-new.php'                => 'https://wordpress.com/people/new/%home_url%',
+		'wp-admin/post-new.php'  => '/wp-admin/post-new.php?post_type=post&calypsoify=1',
+		'wp-admin/media-new.php' => 'https://wordpress.com/media/%home_url%',
+		'wp-admin/user-new.php'  => 'https://wordpress.com/people/new/%home_url%',
+	);
+
+	/**
+	 * Constructor.
+	 */
+	public function __construct() {
+		if ( ! wpcom_should_disable_calypso_links( 'edit.php?post_type=page' ) ) {
+			$this->map_wp_admin_url_to_calypso_url['wp-admin/post-new.php?post_type=page'] = 'https://wordpress.com/page/%home_url%';
+		}
 
 		/**
 		 * Jetpack
 		 */
-		'wp-admin/post-new.php?post_type=jetpack-testimonial' => 'https://wordpress.com/types/jetpack-testimonial/%home_url%',
-		'wp-admin/post-new.php?post_type=jetpack-portfolio' => 'https://wordpress.com/types/jetpack-portfolio/%home_url%',
-	);
+		if ( ! wpcom_should_disable_calypso_links( 'edit.php?post_type=jetpack-testimonial' ) ) {
+			$this->map_wp_admin_url_to_calypso_url['wp-admin/post-new.php?post_type=jetpack-testimonial'] = 'https://wordpress.com/types/jetpack-testimonial/%home_url%';
+		}
+
+		if ( ! wpcom_should_disable_calypso_links( 'edit.php?post_type=jetpack-portfolio' ) ) {
+			$this->map_wp_admin_url_to_calypso_url['wp-admin/post-new.php?post_type=jetpack-portfolio'] = 'https://wordpress.com/types/jetpack-portfolio/%home_url%';
+		}
+	}
 
 	/**
 	 * Adds a node to the menu.