+++ projects/packages/my-jetpack/src/class-rest-products.php
@@ -24,7 +24,7 @@ class REST_Products {
 				array(
 					'methods'             => \WP_REST_Server::READABLE,
 					'callback'            => __CLASS__ . '::get_products',
-					'permission_callback' => __CLASS__ . '::permissions_callback',
+					'permission_callback' => __CLASS__ . '::view_products_permissions_callback',
 					'args'                => array(
 						'products' => array(
 							'description'       => __( 'Comma seperated list of product slugs that should be retrieved.', 'jetpack-my-jetpack' ),
@@ -101,7 +101,7 @@ class REST_Products {
 				array(
 					'methods'             => \WP_REST_Server::READABLE,
 					'callback'            => __CLASS__ . '::get_products_by_ownership',
-					'permission_callback' => __CLASS__ . '::permissions_callback',
+					'permission_callback' => __CLASS__ . '::view_products_permissions_callback',
 				),
 			)
 		);
@@ -131,6 +131,15 @@ class REST_Products {
 	 */
 	public static function permissions_callback() {
 		return current_user_can( 'manage_options' );
+	}
+
+	/**
+	 * Check if the user is permitted to view the product and product info
+	 *
+	 * @return bool
+	 */
+	public static function view_products_permissions_callback() {
+		return current_user_can( 'edit_posts' );
 	}
 
 	/**