+++ projects/js-packages/babel-plugin-replace-textdomain/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.43] - 2025-02-24
+### Changed
+- Update dependencies.
+
 ## [1.0.42] - 2024-12-16
 ### Changed
 - Internal updates.
@@ -189,6 +193,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release.
 - Replace missing domains too.
 
+[1.0.43]: https://github.com/Automattic/babel-plugin-replace-textdomain/compare/v1.0.42...v1.0.43
 [1.0.42]: https://github.com/Automattic/babel-plugin-replace-textdomain/compare/v1.0.41...v1.0.42
 [1.0.41]: https://github.com/Automattic/babel-plugin-replace-textdomain/compare/v1.0.40...v1.0.41
 [1.0.40]: https://github.com/Automattic/babel-plugin-replace-textdomain/compare/v1.0.39...v1.0.40