+++ projects/js-packages/charts/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2] - 2025-02-19
+### Added
+- Add yScale options support. [#41866]
+
+### Changed
+- Updated package dependencies. [#41578]
+
+### Fixed
+- Line chart: support custom tooltips [#41738]
+
 ## [0.8.1] - 2025-02-04
 ### Fixed
 - Charts: fixed type exports [#41562]
@@ -109,6 +119,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed lints following ESLint rule changes for TS [#40584]
 - Fixing a bug in Chart storybook data. [#40640]
 
+[0.8.2]: https://github.com/Automattic/charts/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/Automattic/charts/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/Automattic/charts/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/Automattic/charts/compare/v0.7.0...v0.7.1