+++ projects/packages/test-environment/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.2 - 2025-02-24
+### Changed
+- Update dependencies.
+
 ## 0.1.1 - 2025-02-17
 ### Changed
 - Internal updates.